requirements/production.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/production.in
#
amqp==5.1.1
    # via kombu
arrow==1.2.3
    # via -r requirements/production.in
asgiref==3.7.2
    # via django
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.2
    # via
    #   django-bootstrap-v5
    #   html-sanitizer
billiard==4.1.0
    # via celery
boto3==1.28.21
    # via django-storages
botocore==1.31.21
    # via
    #   boto3
    #   s3transfer
celery[redis]==5.3.1
    # via
    #   -r requirements/production.in
    #   django-celery-beat
certifi==2023.7.22
    # via sentry-sdk
cffi==1.15.1
    # via cryptography
click==8.1.6
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
crispy-bootstrap5==0.7
    # via -r requirements/production.in
cron-descriptor==1.4.0
    # via django-celery-beat
cryptography==41.0.3
    # via django-rest-knox
django==4.2.4
    # via
    #   -r requirements/production.in
    #   crispy-bootstrap5
    #   django-appconf
    #   django-bootstrap-v5
    #   django-celery-beat
    #   django-crispy-forms
    #   django-extensions
    #   django-filter
    #   django-health-check
    #   django-redis
    #   django-rest-knox
    #   django-storages
    #   django-timezone-field
    #   djangorestframework
    #   drf-spectacular
    #   drf-standardized-errors
django-appconf==1.0.5
    # via django-imagekit
django-bootstrap-v5==1.0.11
    # via -r requirements/production.in
django-celery-beat==2.5.0
    # via -r requirements/production.in
django-crispy-forms==2.0
    # via
    #   -r requirements/production.in
    #   crispy-bootstrap5
django-environ==0.10.0
    # via -r requirements/production.in
django-extensions==3.2.3
    # via -r requirements/production.in
django-filter==23.2
    # via -r requirements/production.in
django-health-check==3.17.0
    # via -r requirements/production.in
django-imagekit==4.1.0
    # via -r requirements/production.in
django-redis==5.3.0
    # via -r requirements/production.in
django-rest-knox==4.2.0
    # via -r requirements/production.in
django-storages[boto3]==1.13.2
    # via -r requirements/production.in
django-timezone-field==5.1
    # via django-celery-beat
djangorestframework==3.14.0
    # via
    #   -r requirements/production.in
    #   django-rest-knox
    #   drf-spectacular
    #   drf-standardized-errors
drf-spectacular==0.26.4
    # via
    #   -r requirements/production.in
    #   drf-standardized-errors
drf-standardized-errors[openapi]==0.12.5
    # via -r requirements/production.in
execnet==2.0.2
    # via pytest-xdist
factory-boy==3.3.0
    # via -r requirements/production.in
faker==19.3.0
    # via factory-boy
html-sanitizer==2.2.0
    # via -r requirements/production.in
inflection==0.5.1
    # via
    #   drf-spectacular
    #   drf-standardized-errors
iniconfig==2.0.0
    # via pytest
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.19.0
    # via drf-spectacular
jsonschema-specifications==2023.7.1
    # via jsonschema
kombu==5.3.1
    # via celery
lxml==4.9.3
    # via html-sanitizer
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mistune==3.0.1
    # via -r requirements/production.in
packaging==23.1
    # via
    #   pytest
    #   pytest-sugar
pilkit==2.0
    # via django-imagekit
pillow==9.5.0
    # via -r requirements/production.in
pluggy==1.2.0
    # via pytest
prompt-toolkit==3.0.39
    # via click-repl
psycopg[binary]==3.1.10
    # via -r requirements/production.in
psycopg-binary==3.1.10
    # via psycopg
pycparser==2.21
    # via cffi
pygments==2.16.1
    # via rich
pytest==7.4.0
    # via
    #   -r requirements/production.in
    #   pytest-django
    #   pytest-sugar
    #   pytest-xdist
pytest-django==4.5.2
    # via -r requirements/production.in
pytest-sugar==0.9.7
    # via -r requirements/production.in
pytest-xdist==3.3.1
    # via -r requirements/production.in
python-crontab==3.0.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
    #   celery
    #   faker
    #   python-crontab
pytz==2023.3
    # via
    #   django-timezone-field
    #   djangorestframework
pyuwsgi==2.0.21
    # via -r requirements/production.in
pyyaml==6.0.1
    # via drf-spectacular
redis==4.6.0
    # via
    #   celery
    #   django-redis
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rich==13.5.2
    # via -r requirements/production.in
rpds-py==0.9.2
    # via
    #   jsonschema
    #   referencing
s3transfer==0.6.1
    # via boto3
sentry-sdk==1.29.2
    # via -r requirements/production.in
six==1.16.0
    # via
    #   django-imagekit
    #   python-dateutil
soupsieve==2.4.1
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
termcolor==2.3.0
    # via pytest-sugar
typing-extensions==4.7.1
    # via psycopg
tzdata==2023.3
    # via
    #   celery
    #   django-celery-beat
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.16
    # via
    #   botocore
    #   sentry-sdk
uwsgitop==0.11
    # via -r requirements/production.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
